refactor: Update App.ts to import and call setThemeTemplate function
